requirements.txt
opencv-python
numpy
matplotlib
Pillow >= 8.0